requirements.txt
attrs==23.1.0
blinker==1.7.0
click==8.1.7
colorama==0.4.6
flasgger==0.9.7.1
Flask==3.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.20.0
jsonschema-specifications==2023.11.1
MarkupSafe==2.1.3
mistune==3.0.2
packaging==23.2
Pillow==10.1.0
PyYAML==6.0.1
referencing==0.31.0
rpds-py==0.13.1
six==1.16.0
Werkzeug==3.0.1
